refactor: move converged charging request validator to util
